@@ -6,8 +6,13 @@ log = get_logger()
 
 # Does nothing currently
 from mailbag.derivative import Derivative
+
+
 class ExampleDerivative(Derivative):
-    derivative_name = 'example'
+    derivative_name = "example"
+    derivative_format = "example"
+    derivative_agent = ""
+    derivative_agent_version = ""
 
     def __init__(self, email_account, **kwargs):
         log.debug("Setup account")
