@@ -1,8 +1,10 @@
 from os import listdir
-from os.path import  basename, dirname, exists, isfile, join
+from os.path import basename, dirname, exists, isfile, join
 from importlib.machinery import SourceFileLoader
 
 from abc import ABC, abstractmethod
+
+
 class EmailAccount(ABC):
     """EmailAccount - abstract base class and registry for concrete email format parsers
 
@@ -19,9 +21,11 @@ class EmailAccount(ABC):
     registry = {}
 
     def __init_subclass__(cls, **kwargs):
-        """Enforce format_name attribute on subclasses, register them"""
-        if not hasattr(cls, 'format_name'):
-            raise RuntimeError("EmailAccount subclass must have `format_name` attribute")
+        """Enforce format descriptive attributes on subclasses, register them"""
+        format_attrs = ["format_name", "format_agent", "format_agent_version"]
+        for attr in format_attrs:
+            if not hasattr(cls, attr):
+                raise RuntimeError("EmailAccount subclass must have `" + attr + "` attribute")
 
         super().__init_subclass__(**kwargs)
         __class__.registry[cls.format_name] = cls
@@ -29,7 +33,7 @@ class EmailAccount(ABC):
     @abstractmethod
     def __init__(self, target_account, args, **kwargs):
         """Implement a method that finds and sets up an account so that `account_data` and `messages` can
-be called on it."""
+        be called on it."""
         pass
 
     @abstractmethod
@@ -42,22 +46,22 @@ be called on it."""
         """Generator method that yields one message at a time from an email account as `Email`s."""
         pass
 
+
 def import_formats(additional_dirs=None):
     if not additional_dirs:
         additional_dirs = []
 
-    dirs = [join(dirname(__file__), 'formats'), *additional_dirs]
+    dirs = [join(dirname(__file__), "formats"), *additional_dirs]
 
     for formats_dir in reversed(dirs):
-        if not exists(formats_dir): continue
+        if not exists(formats_dir):
+            continue
 
         for filename in listdir(formats_dir):
             module = basename(filename)[:-3]
             full_path = join(formats_dir, filename)
             # skip if not a normal, non underscored file ending in .py
-            if module.startswith('_') or \
-               not isfile(full_path) or \
-               filename[-3:] != '.py':
+            if module.startswith("_") or not isfile(full_path) or filename[-3:] != ".py":
                 continue
 
             SourceFileLoader(module, full_path).load_module()
