@@ -4,8 +4,13 @@
 # Does nothing currently
 from mailbag.email_account import EmailAccount
 from mailbag.models import Email
+
+
 class ExampleAccount(EmailAccount):
-    format_name = 'example'
+    format_name = "example"
+    format_details = ""
+    format_agent = ""
+    format_agent_version = ""
 
     def __init__(self, target_account, args, **kwargs):
         print("Parsity parse")
